chore: rearrange helpers
